feat: make IAddonManager
