+++ Sharp.Core/Managers/SharedManager.cs
@@ -85,4 +85,7 @@ internal class SharedManager : ISharedManager
 
     public IParticleManager GetParticleManager()
         => _serviceProvider.GetRequiredService<ICoreParticleManager>();
+
+    public IAddonManager GetAddonManager()
+        => _serviceProvider.GetRequiredService<ICoreAddonManager>();
 }
