+++ Sharp.Shared/ISharedSystem.cs
@@ -55,4 +55,6 @@ public interface ISharedSystem
     ISharpModuleManager GetSharpModuleManager();
 
     IParticleManager GetParticleManager();
+
+    IAddonManager GetAddonManager();
 }
